refactor: move tests to expected pacakage
