+++ src/test/java/io/pakland/mdas/githubstats/domain/entity/CSVTest.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain;
+package io.pakland.mdas.githubstats.domain.entity;
 
 import java.util.List;
