+++ src/test/java/io/pakland/mdas/githubstats/domain/entity/UserTest.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain;
+package io.pakland.mdas.githubstats.domain.entity;
 
 import org.junit.jupiter.api.Test;
